+++ include/ellis/stream/fd_output_stream.hpp
@@ -24,8 +24,8 @@ class fd_output_stream : public sync_output_stream {
   int m_valid = 0;
 public:
   fd_output_stream(int fd);
-  bool next_output_buf(byte **buf, int *bytecount) override;
-  bool emit(int bytecount) override;
+  bool next_output_buf(byte **buf, size_t *bytecount) override;
+  bool emit(size_t bytecount) override;
   std::unique_ptr<err> get_output_error() override;
 };
