+++ include/ellis/stream/tcp_client_stream.hpp
@@ -26,15 +26,15 @@ public:
 
   ~tcp_stream();
 
-  bool next_input_buf(byte **buf, int *bytecount) override;
+  bool next_input_buf(byte **buf, size_t *bytecount) override;
 
-  void put_back(int bytecount) override;
+  void put_back(size_t bytecount) override;
 
   std::unique_ptr<err> get_input_error() override;
 
-  bool next_output_buf(byte **buf, int *bytecount) override;
+  bool next_output_buf(byte **buf, size_t *bytecount) override;
 
-  bool emit(int bytecount) override;
+  bool emit(size_t bytecount) override;
 
   std::unique_ptr<err> get_output_error() override;
 };
