+++ include/ellis/sync_input_stream.hpp
@@ -30,10 +30,10 @@ public:
    *   true  --> Data is available (and has been returned).
    *   false --> Data is unavailable (check get_input_error() for details).
    */
-  virtual bool next_input_buf(byte **buf, int *bytecount) = 0;
+  virtual bool next_input_buf(byte **buf, size_t *bytecount) = 0;
 
   /** Put back data at the end of the block last returned from next_block(). */
-  virtual void put_back(int bytecount) = 0;
+  virtual void put_back(size_t bytecount) = 0;
 
   /** Return the error details.  Caller owns it now.  */
   virtual std::unique_ptr<err> get_input_error() = 0;
